+++ detector/YOLOV5/detector.py
@@ -62,7 +62,7 @@ class YOLOv5(object):
         img = letterbox(ori_img, new_shape=self.size)[0]
         img = img[:, :, ::-1].transpose(2, 0, 1)  # BGR to RGB, to 3x416x416
         img = np.ascontiguousarray(img)        
-        img = img.astype(np.float) / 255.
+        img = img.astype(np.float32) / 255.
         img = torch.from_numpy(img).float().unsqueeze(0)
         # forward
         with torch.no_grad():
