+++ detector/YOLOv3/detector.py
@@ -37,7 +37,7 @@ class YOLOv3(object):
     def __call__(self, ori_img):
         # img to tensor
         assert isinstance(ori_img, np.ndarray), "input must be a numpy array!"
-        img = ori_img.astype(np.float) / 255.
+        img = ori_img.astype(np.float32) / 255.
 
         img = cv2.resize(img, self.size)
         img = torch.from_numpy(img).float().permute(2, 0, 1).unsqueeze(0)
